Main window: Remove global proxy style
- That was interfering with displaying the right padding for menu items
(don't know why).
- Move the two styles declared on it to their respective widgets. This
also fixes setting up the right icon size for context menus, which was
not working.
- Simplify applying stylesheet to the main window.
